pyproject.toml
[project]
name = "rl-coach-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-ai>=1.66.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
